op-cancel-appointment - refactor dog-gender and define appointment-status
