+++ src/app/user/(logged-in)/my-pets/_components/general-dog-form.tsx
@@ -10,11 +10,8 @@ import {
   BarkFormRadioGroup,
 } from "@/components/bark/bark-form";
 import { isValidWeightKg } from "@/lib/utilities/bark-utils";
-import {
-  DOG_ANTIGEN_PRESENCE,
-  DOG_GENDER,
-  YES_NO_UNKNOWN,
-} from "@/lib/data/db-enums";
+import { DOG_ANTIGEN_PRESENCE, YES_NO_UNKNOWN } from "@/lib/data/db-enums";
+import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
