+++ src/lib/data/db-enums.ts
@@ -94,14 +94,6 @@ export const YES_NO_UNKNOWN = {
 
 export type YesNoUnknown = ObjectValues<typeof YES_NO_UNKNOWN>;
 
-export const DOG_GENDER = {
-  MALE: "MALE",
-  FEMALE: "FEMALE",
-  UNKNOWN: "UNKNOWN",
-} as const;
-
-export type DogGender = ObjectValues<typeof DOG_GENDER>;
-
 export const DOG_ANTIGEN_PRESENCE = {
   POSITIVE: "POSITIVE",
   NEGATIVE: "NEGATIVE",
