+++ tests/_fixtures.ts
@@ -22,12 +22,11 @@ import {
 } from "@/lib/data/db-models";
 import {
   DOG_ANTIGEN_PRESENCE,
-  DOG_GENDER,
   DogAntigenPresence,
-  DogGender,
   YES_NO_UNKNOWN,
   YesNoUnknown,
 } from "@/lib/data/db-enums";
+import { DOG_GENDER, DogGender } from "@/lib/bark/models/dog-gender";
 import { USER_RESIDENCY } from "@/lib/data/db-enums";
 import { dbInsertAdmin, dbSelectAdmin } from "@/lib/data/db-admins";
 import { Pool } from "pg";
