+++ tests/admin/get-incomplete-profiles.test.ts
@@ -9,11 +9,8 @@ import {
 } from "../_fixtures";
 import { getIncompleteProfiles } from "@/lib/admin/actions/get-incomplete-profiles";
 import { Pool } from "pg";
-import {
-  DOG_GENDER,
-  PARTICIPATION_STATUS,
-  YES_NO_UNKNOWN,
-} from "@/lib/data/db-enums";
+import { PARTICIPATION_STATUS, YES_NO_UNKNOWN } from "@/lib/data/db-enums";
+import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
 import {
   DEFAULT_DATE_TIME_FORMAT,
   SINGAPORE_TIME_ZONE,
